modify group id and artifact id
