--- src/main/java/com/bjit/mailservice/LibEmailServiceApplication.java
@@ -1,15 +1,0 @@
-package com.bjit.mailservice;
-
-import jakarta.mail.MessagingException;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * @hidden
- */
-@SpringBootApplication
-public class LibEmailServiceApplication {
-    public static void main(String[] args) throws MessagingException {
-        SpringApplication.run(LibEmailServiceApplication.class, args);
-    }
-}
