--- src/main/java/com/bjit/mailservice/exception/EmailException.java
@@ -1,14 +1,0 @@
-/**
- * @author Khalid
- * Date: 1/25/2024
- * Time: 4:06 PM
- * Project Name: lib-email-service
- */
-
-package com.bjit.mailservice.exception;
-
-public class EmailException extends RuntimeException {
-    public EmailException(String message) {
-        super(message);
-    }
-}
