+++ src/main/java/com/bjitgroup/emailservice/exception/EmailException.java
@@ -1,0 +1,14 @@
+/**
+ * @author Khalid
+ * Date: 1/25/2024
+ * Time: 4:06 PM
+ * Project Name: lib-email-service
+ */
+
+package com.bjitgroup.emailservice.exception;
+
+public class EmailException extends RuntimeException {
+    public EmailException(String message) {
+        super(message);
+    }
+}
