--- src/test/java/com/bjit/mailservice/LibEmailServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bjit.mailservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LibEmailServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
