+++ src/test/java/com/bjitgroup/emailservice/LibEmailServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bjitgroup.emailservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LibEmailServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
